Change Sublet Amenity Return Formatting (#251)
* Sublet Image Handling

* Fix linting whoops

* Add list comprehensions + prevalidation to images

* Fix misnamed serializer field

* Fix test cases

* Add separate routes for image creation and deletion

* Begin work on sublet testing

* Mocked AWS call

* Add tests for multiple images + deletion

* How did this not fail in previous runs haha

* Add sublet negotiable, swap max/min to reg price fields

* Fix filtering

* Change amenities return formatting + add comments for sublet/urls

* Alter baths to be 1-point decimal

* Fix testing for decimal baths + add testing for new amenities formatting

---------

+++ backend/sublet/models.py
@@ -39,7 +39,7 @@ class Sublet(models.Model):
     title = models.CharField(max_length=255)
     address = models.CharField(max_length=255, null=True, blank=True)
     beds = models.IntegerField(null=True, blank=True)
-    baths = models.IntegerField(null=True, blank=True)
+    baths = models.DecimalField(max_digits=3, decimal_places=1, null=True, blank=True)
     description = models.TextField(null=True, blank=True)
     external_link = models.URLField(max_length=255)
     price = models.IntegerField()
